requirements.txt
python-dotenv~=0.20.0
setuptools~=57.0.0
pandas~=1.3.5
requests~=2.28.1
OSMPythonTools~=0.3.4
Pillow~=9.1.0
matplotlib~=3.5.1
networkx~=2.6.3
numpy~=1.21.5